+++ aimmo-game/simulation/event.py
@@ -1,12 +1,8 @@
 from collections import namedtuple
 
-ReceivedAttackEvent = namedtuple(
-    "ReceivedAttackEvent", ["attacking_avatar", "damage_dealt"]
-)
+ReceivedAttackEvent = namedtuple("ReceivedAttackEvent", ["attacking_avatar", "damage_dealt"])
 
-PerformedAttackEvent = namedtuple(
-    "PerformedAttackEvent", ["attacked_avatar", "target_location", "damage_dealt"]
-)
+PerformedAttackEvent = namedtuple("PerformedAttackEvent", ["attacked_avatar", "target_location", "damage_dealt"])
 
 FailedAttackEvent = namedtuple("FailedAttackEvent", ["target_location"])
 
@@ -17,3 +13,7 @@ FailedMoveEvent = namedtuple("FailedMoveEvent", ["source_location", "target_loca
 PickedUpEvent = namedtuple("PickedUpEvent", ["interactable"])
 
 FailedPickupEvent = namedtuple("FailedPickupEvent", [])
+
+DroppedEvent = namedtuple("DroppedEvent", ["index"])
+
+FailedDropEvent = namedtuple("FailedDropEvent", [])