+++ aimmo-game/simulation/game_state.py
@@ -28,6 +28,7 @@ class GameState:
     def serialize(self):
         return {
             "era": self.worksheet.era,
+            "worksheetID": self.worksheet.worksheet_id,
             "southWestCorner": self.world_map.get_serialized_south_west_corner(),
             "northEastCorner": self.world_map.get_serialized_north_east_corner(),
             "players": self.avatar_manager.serialize_players(),