+++ game_frontend/src/pyodide/pyodideRunner.ts
@@ -6,17 +6,26 @@ import { PyodideWorker } from './webWorker'
 let worker: ModuleThread<PyodideWorker>
 let workerReady = false
 
-export async function initializePyodide () {
+export async function initializePyodide() {
   await initializePyodideWorker()
   workerReady = true
 }
 
-async function initializePyodideWorker () {
+async function initializePyodideWorker() {
   worker = await spawn<PyodideWorker>(new Worker('./webWorker.ts'))
   await worker.initializePyodide()
 }
 
-export async function updateAvatarCode (
+export async function checkIfBadgeEarned(
+  badges: string,
+  result: ComputedTurnResult,
+  userCode: string,
+  gameState: any
+): Promise<string> {
+  return worker.checkIfBadgeEarned(badges, result, userCode, gameState)
+}
+
+export async function updateAvatarCode(
   userCode: string,
   gameState: any,
   playerAvatarID: number = 0
@@ -28,7 +37,7 @@ export async function updateAvatarCode (
   )
 }
 
-export async function resetWorker (userCode: string, playerAvatarID: number) {
+export async function resetWorker(userCode: string, playerAvatarID: number) {
   workerReady = false
   await Thread.terminate(worker)
   await initializePyodideWorker()
@@ -36,7 +45,7 @@ export async function resetWorker (userCode: string, playerAvatarID: number) {
   workerReady = true
 }
 
-async function runIfWorkerReady (
+async function runIfWorkerReady(
   func: () => Promise<ComputedTurnResult>,
   turnCount: number
 ): Promise<ComputedTurnResult> {
@@ -46,7 +55,7 @@ async function runIfWorkerReady (
     return Promise.resolve({
       action: { action: { action_type: 'wait' } },
       log: '',
-      turnCount: turnCount + 1
+      turnCount: turnCount + 1,
     })
   }
 }