+++ game_frontend/src/pyodide/webWorker.ts
@@ -1,15 +1,17 @@
 /* eslint-env worker */
 import { expose } from 'threads/worker'
+import { checkIfBadgeEarned, filterByWorksheet } from './badges'
 import ComputedTurnResult from './computedTurnResult'
 
+let pyodide: Pyodide
+
 function getAvatarStateFromGameState(gameState: any, playerAvatarID: number): object {
-  return gameState.players.find(player => player.id === playerAvatarID)
+  return gameState.players.find((player) => player.id === playerAvatarID)
 }
 
 async function initializePyodide() {
-  self.languagePluginUrl = 'https://pyodide-cdn2.iodide.io/v0.15.0/full/'
-  importScripts('https://pyodide-cdn2.iodide.io/v0.15.0/full/pyodide.js')
-  await languagePluginLoader
+  importScripts('https://cdn.jsdelivr.net/pyodide/v0.20.0/full/pyodide.js')
+  pyodide = await loadPyodide()
   await pyodide.loadPackage(['micropip'])
   await pyodide.runPythonAsync(`
 import micropip
@@ -18,13 +20,17 @@ micropip.install("${self.location.origin}/static/worker/aimmo_game_worker-0.0.0-
   `)
 
   await pyodide.runPythonAsync(`
-from simulation import direction
-from simulation import location
-from simulation.action import MoveAction, PickupAction, WaitAction, MoveTowardsAction
-from simulation.world_map import WorldMapCreator
-from simulation.avatar_state import create_avatar_state
-from io import StringIO
 import contextlib
+import sys
+
+from js import Object
+from io import StringIO
+from pyodide import to_js
+
+from simulation import direction, location
+from simulation.action import MoveAction, PickupAction, WaitAction, MoveTowardsAction, DropAction
+from simulation.avatar_state import create_avatar_state
+from simulation.world_map import WorldMapCreator
 
 
 @contextlib.contextmanager
@@ -61,13 +67,13 @@ with capture_output() as output:
     serialized_action = action.serialise()
 stdout, stderr = output
 logs = stdout.getvalue() + stderr.getvalue()
-{"action": serialized_action, "log": logs, "turnCount": game_state["turnCount"] + 1}
+to_js({"action": serialized_action, "log": logs, "turnCount": game_state["turnCount"] + 1}, dict_converter=Object.fromEntries)
     `)
   } catch (error) {
     return Promise.resolve({
       action: { action_type: 'wait' },
       log: simplifyErrorMessageInLog(error.toString()),
-      turnCount: gameState.turnCount + 1
+      turnCount: gameState.turnCount + 1,
     })
   }
 }
@@ -81,10 +87,7 @@ export function simplifyErrorMessageInLog(log: string): string {
     return `Uh oh! Something isn't correct on line ${matches[1]}. Here's the error we got:\n${simpleError}`
   }
   // error not in next_turn function
-  return log
-    .split('\n')
-    .slice(-2)
-    .join('\n')
+  return log.split('\n').slice(-2).join('\n')
 }
 
 export async function updateAvatarCode(
@@ -105,14 +108,14 @@ export async function updateAvatarCode(
     return Promise.resolve({
       action: { action_type: 'wait' },
       log: '',
-      turnCount: turnCount
+      turnCount: turnCount,
     })
   } catch (error) {
     await setAvatarCodeToWaitActionOnError()
     return Promise.resolve({
       action: { action_type: 'wait' },
       log: simplifyErrorMessageInLog(error.toString()),
-      turnCount: turnCount
+      turnCount: turnCount,
     })
   }
 }
@@ -127,7 +130,9 @@ async function setAvatarCodeToWaitActionOnError() {
 const pyodideWorker = {
   initializePyodide,
   computeNextAction,
-  updateAvatarCode
+  updateAvatarCode,
+  checkIfBadgeEarned,
+  filterByWorksheet,
 }
 
 export type PyodideWorker = typeof pyodideWorker