+++ game_frontend/src/redux/features/AvatarWorker/epics.js
@@ -1,4 +1,4 @@
-import { switchMap, mapTo, zip, take, tap, map } from 'rxjs/operators'
+import { switchMap, mapTo, zip, take, tap, map, mergeMap, catchError } from 'rxjs/operators'
 
 import actions from './actions'
 import types from './types'
@@ -6,7 +6,7 @@ import { timeoutIfWorkerTakesTooLong } from './operators'
 import { ofType } from 'redux-observable'
 import { gameTypes } from '../Game'
 import { editorTypes } from '../Editor'
-import { from, Scheduler } from 'rxjs'
+import { from, of, Scheduler } from 'rxjs'
 
 const backgroundScheduler = Scheduler.async
 
@@ -83,9 +83,114 @@ const computeNextActionEpic = (
     )
   )
 
+/**
+ * Get the user's badges information.
+ * @returns a redux action that contains a string storing the user's earned badges information.
+ */
+const getBadgesEpic = (action$, state$, { api }) =>
+  action$.pipe(
+    ofType(types.GET_BADGES_REQUEST),
+    switchMap(() =>
+      action$.pipe(
+        ofType(types.PYODIDE_INITIALIZED),
+        mergeMap((action) =>
+          api.get(`badges/${state$.value.game.connectionParameters.game_id}/`).pipe(
+            map((response) => actions.getBadgesReceived(response.badges)),
+            catchError((error) =>
+              of({
+                type: types.GET_BADGES_FAILURE,
+                payload: error.xhr.response,
+                error: true,
+              })
+            )
+          )
+        )
+      )
+    )
+  )
+
+/**
+ * Whenever the avatar's code is updated, get the user's badges information.
+ * @returns a redux action that contains a string storing the user's earned badges information.
+ */
+const checkBadgesEpic = (action$, state$, { api }) =>
+  action$.pipe(
+    ofType(types.AVATAR_CODE_UPDATED),
+    mergeMap((action) =>
+      api.get(`badges/${state$.value.game.connectionParameters.game_id}/`).pipe(
+        map((response) => actions.checkBadgesReceived(response.badges)),
+        catchError((error) =>
+          of({
+            type: types.GET_BADGES_FAILURE,
+            payload: error.xhr.response,
+            error: true,
+          })
+        )
+      )
+    )
+  )
+
+/**
+ * Every time the avatar's code's been updated and the user's current badges have been read from Django, run a badge
+ * check that evaluates if any new badges have been earned.
+ * @returns a redux action that holds the result of the badge check.
+ */
+const checkBadgesEarnedEpic = (action$, state$, { pyodideRunner: { checkIfBadgeEarned } }) =>
+  action$.pipe(
+    ofType(types.AVATAR_CODE_UPDATED),
+    switchMap(({ payload: computedTurnResult }) =>
+      action$.pipe(
+        ofType(types.BADGES_CHECKED_SUCCESS),
+        switchMap(({ payload: badges }) =>
+          from(
+            checkIfBadgeEarned(
+              badges,
+              computedTurnResult,
+              state$.value.editor.code.codeOnServer,
+              state$.value.game.gameState
+            )
+          )
+        ),
+        map(actions.badgesEarned),
+        catchError((error) =>
+          of({
+            type: types.BADGES_CHECKED_FAILURE,
+            payload: error,
+            error: true,
+          })
+        )
+      )
+    )
+  )
+
+/**
+ * Once the badge check has been run on the user's new code to see if they've earned any new badges, send a POST request
+ * with the updated information.
+ * @returns a redux action that confirms the success or failure of the POST request.
+ */
+const postBadgesEpic = (action$, state$, { api }) =>
+  action$.pipe(
+    ofType(types.BADGES_EARNED),
+    api.post(`/kurono/api/badges/${state$.value.game.connectionParameters.game_id}/`, (action) => {
+      return { badges: action.payload }
+    }),
+    map(() => actions.postBadgesReceived()),
+    catchError((error) =>
+      of({
+        type: types.POST_BADGES_FAILURE,
+        payload: error.xhr.response,
+        error: true,
+      })
+    )
+  )
+
 export default {
   initializePyodideEpic,
   initialUpdateAvatarCodeEpic,
   updateAvatarCodeEpic,
-  computeNextActionEpic
+  computeNextActionEpic,
+  getBadgesEpic,
+  checkBadgesEpic,
+  postBadgesEpic,
+  checkBadgesEarnedEpic,
 }